# perf: teach Read Buffer to materialise string column results in dictionary format