--- a/read_buffer/benches/read.rs
+++ b/read_buffer/benches/read.rs
@@ -4,5 +4,5 @@ use criterion::{criterion_group, criterion_main};
 
 use read_group::read_group;
 
-criterion_group!(benches, read_group,);
+criterion_group!(benches, read_group);
 criterion_main!(benches);