--- a/read_buffer/benches/read.rs
+++ b/read_buffer/benches/read.rs
@@ -1,0 +1,8 @@
+mod read_group;
+
+use criterion::{criterion_group, criterion_main};
+
+use read_group::read_group;
+
+criterion_group!(benches, read_group,);
+criterion_main!(benches);