This PR contributes to our ongoing Query Performance initiative by significantly reducing the amount of memory necessary to create a record batch of results from the Read Buffer, when string columns are involved.

## Rationale

Previously @tustvold taught IOx how to use Arrow Dictionary array types when feeding results to DataFusion and when storing snapshots of data. 

In order for record batch schemas emitted from the Read Buffer to be compatible with other record batches emitted from the MUB it was necessary for the Read Buffer to emit Dictionary arrays for "tag" columns. @tustvold implemented this compatibility layer in https://github.com/influxdata/influxdb_iox/pull/1262. However, there was some non-trivial amount of work needed to teach all the compressed column encodings in the Read Buffer to materialise their results as dictionaries. So in #1262 we just materialised all the string values for a column's results and then squeezed it back into an Arrow dictionary.

This PR effectively does the extra work to teach the column encodings for string columns to only materialise their results as dictionaries.

## Relevant Notes

As it stands today, the IOx schema specifies that columns with the InfluxDB Tag semantic type should be represented as Dictionary Arrow arrays, but String _fields_ are represented as `StringArray`. 

https://github.com/influxdata/influxdb_iox/blob/9b42c1a065aab8271eae3f7fbf44c166b587b005/internal_types/src/schema.rs#L651-L660

Because the Read Buffer's columnar encodings do not have a concept of semantic types (their level of abstraction is lower than that) I decided to teach them to materialise their values as either a dictionary format or an array format and then higher up I decided which of those paths to choose based on the semantic type of the column.

If we decide to encode influx string field columns as dictionaries (FWIW - I don't think this is a high priority right now) then it would only be a small change to the `read_filter` implementation in the Read Buffer to support that.

## Performance

I added benchmarks to exercise the `read_filter` implementation in the Read Buffer, focussing on two situations:

 - no predicate with different numbers of projected columns.
 - with a predicate on a column but projecting all columns.

The performance of the "no predicate" case is directly related to what the Chunk Merger (https://github.com/influxdata/influxdb_iox/pull/1517) will be calling, so wins there will help performance when merging chunks.

Further, since IOx currently hasn't wired up predicate pushdown to the Read Buffer's implementation all queries effectively ask for all the data on every query...

Here is a flamegraph showing that materialisation of string columns in the Read Buffer can cause significant CPU utilisation. This should be significantly reduced when this PR lands:

<img width="1552" alt="Screenshot 2021-05-19 at 15 57 02" src="https://user-images.githubusercontent.com/501993/118845826-044a2180-b8c4-11eb-96e5-d56a715440be.png">


### Benchmark Results

```
⇒  critcmp main pr
group                                                 main                                     pr
-----                                                 ----                                     --
read_filter/no_pred/cardinality_200000_rows_200000    9.46     10.6±0.10ms 17.9 MElem/sec      1.00  1124.5±71.39µs 169.6 MElem/sec
read_filter/no_pred/cardinality_2000_rows_200000      13.93     6.9±0.03ms 27.8 MElem/sec      1.00    492.9±2.44µs 387.0 MElem/sec
read_filter/no_pred/cardinality_200_rows_200000       13.59     6.4±0.06ms 30.0 MElem/sec      1.00    468.6±1.42µs 407.0 MElem/sec
read_filter/no_pred/cardinality_2_rows_200000         12.87     6.0±0.10ms 31.8 MElem/sec      1.00    465.7±1.06µs 409.6 MElem/sec
read_filter/with_pred/rows_100000                     1.48     52.4±1.04ms 1865.4 KElem/sec    1.00     35.5±0.17ms  2.7 MElem/sec```

For the no predicate case this PR improves `read_filter` performance by `9–13x`. For the case we have predicates we need to do some work to rewrite the encoded representations we put into Arrow's dictionary. However, we still get an improvement of `1.48x`. Naturally we can improve further in the future but this is good for now I think.